+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.2.0] - 2021-??-??
+ - Fixed Texture3DAtlas assuming Texture3D support for compressed formats ([Case 1208832](https://issuetracker.unity3d.com/issues/unable-to-create-a-texture3d-with-a-compressed-format)) works in Unity 2019.4. It works in Unity 2020.1 and newer only. Thanks to Richard for pointing this out.
  - Fixed Texture3DAtlas not updating its texture format when changing the build target with [Asset Import Pipeline V2](https://blogs.unity3d.com/2019/10/31/the-new-asset-import-pipeline-solid-foundation-for-speeding-up-asset-imports/) being used. Thanks to Bastien for the help (actually providing the fix/workaround).
 
 ## [1.1.0] - 2020-10-31